# Cumulativity for inductive types