This PR implements cumulativity for inductive types.

In particular it adds the option ```Ind Cumulatiity``` which can be (globally or locally) set, e.g., ```Global Set Ind Cumulativity```. It can also be enabled or disabled selectively when an defining an inductive type: ```Cumulative Inductive ...``` or ```NonCumulative Inductive ...```.

At the time of definition a subtyping relation is __inferred__ for instances of universe polymorphic cumulative inductive types. This subtyping relation is subsequently used in cumulativity and conversion tests.

For an example of how it works see the file ```test-suite/coqchk/cumulativity.v```.