@JasonGross 2017-05-04
Nice!

Any reason to abbreviate to `Ind` rather then `Inductive `?