@ejgallego 2017-05-04
Congrats Amin on this work. I have an extremely naïve question: is the option really necessary, or could we imagine a mode where all the inductives are cumulative?