@JasonGross 2017-05-05
@ejgallego: what do you mean?  In particular, what makes this a questing to ask of this feature, and not of `Set Universe Polymorphism `?