@maximedenes 2017-05-05
@JasonGross I've always wondered for `Set Universe Polymorphism` too. In fact, when we see some incompatibilities when turning on this option, I still don't know if they are expected or considered a bug.