@amintimany 2017-05-05
@JasonGross the parser can't handle it. I used ```Inductive``` first but then had to change it to ```ind```.