@amintimany 2017-05-05
@ejgallego the analysis to find the subtyping relation can be expensive. It might have to reduce the term and in some cases perhaps down to normal form. In fact at first it was always enabled but then we noticed that finding the subtyping for [https://github.com/amintimany/coq/blob/cum-ind-merging/test-suite/bugs/closed/3330.v#L1078](https://github.com/amintimany/coq/blob/cum-ind-merging/test-suite/bugs/closed/3330.v#L1078) takes minutes.