@maximedenes 2017-05-05
@amintimany This is impressive work! Is there any theoretical justification for this subtyping relation and its compatibility with the usual axioms?