@amintimany 2017-05-05
@maximedenes oh we tried making all the standard library polymorphic and we found out interesting problems with that.

The first problem was how universe polymorphic constructs are handled in modules. In articular, if you have a universe polymorphic declaration in a module type. For instance the following code is rejected at the last line
```
Module Type A.
  Polymorphic Axiom D : Type -> Type.
End A.

Module B : A.
  Polymorphic Definition D := fun X => forall Y Z : Type, X -> Y -> Z -> X.
End B.
```
with error ```Error: Signature components for label D do not match: polymorphic universe instances do not match.``` Because the ```D``` in the module type has two universes while the ```D``` in the module B as three.

And of course, you can't have a parameter in a module type being polymorphic and in the instantiation polymorphic (and vice versa I think).

These problems meant that we had to make almost anything dealing with ```theories/Numbers``` monomorphic which is already a good chunk of the standard library.

Even if we forget about this problem there is another more fatal problem, __the plugins__. Most plugins, like omega, romega and setoid_ring when generating terms expect the standard library definitions, e.g., eq, conj, iff, ... and lemmas about them to be monomorphic and when you turn them polymorphic they just can't work. We (mostly @mattam82) fixed these bugs, although there are probably some (many!?) left.

The last bug though that stopped us from going any further was the __funind__ plugin. This plugin does very complicated things with terms to construct induction principles and this somehow (I don't really understand why or how) needs to type check terms, detype them, do some analysis on them and then retype them, probably in multiple stages even. This means every time new universes are assigned which are different from the old ones. Effectively, a lot of information is lost about universes in this process.

It seems the only way to solve this problem is to reimplement the whole plugin. But I guess that's what's going to happen with Equations. It is supposed to replace funinf if I'm not mistaken.