@amintimany 2017-05-05
@maximedenes the justification is that in the usual __set theoretic__ models we have that whenever a type ```A``` is a subtype of type ```B```, e.g., ```Type@{1} <= Type@{2}``` we have the interpretation of types ```[[A]]``` is a __subset__ of the interpretation of ```[[B]]```. This means that in all such models this property already extends to inductive types (which are just the least fixpoints of closing under the constructors of the inductive type) and in fact such models already satisfy this subtyping property. We looked into it with @mattam82 and at the moment I am trying to write this down more precisely by adapting one of the existing models of pCIC.