@herbelin 2017-05-05
For the "Set Inductive Cumulativity" being blocked (by "Inductive" being a keyword), this is solvable by either renouncing to "Inductive" being a keyword (there were a discussion about 3 weeks ago on Gitter about whether this is really worth to continue having only some key words of the specification language being formally of the "keyword" lexical kind, but not all), or by hard-wiring an explicit rule "Set Inductive" rule in the parser.

In any case, congratulations as well.