@Zimmi48 2017-05-05
FTR here is the beginning of said discussion on Gitter: https://gitter.im/coq/coq?at=58ed380ba0e4856242f54756
and here is the corresponding bug report: https://coq.inria.fr/bugs/show_bug.cgi?id=5464

@herbelin What would we lose by having less keywords? (The other solution you propose seems like a hack to me.)