@amintimany 2017-05-05
@Zimmi48 @herbelin I see. Just for the sake of testing I just turned ```Inductive```, ```CoInductive``` and ```Variant``` into ```IDENT``` and changed the option to ```Set Inductive Cumulativity```. Everything compiled fine and nothing changed for the test suite. I don't see why we have to keep them as keywords!? I can commit that on this PR as well.