@herbelin 2017-05-05
@Zimmi48: Technically, we need articulation words (e.g. `in`, `with`) to be keywords for parsing purpose, but such as the grammar is done, we don't need words starting commands to be keywords.

So, whether `Inductive` should be declared as a keyword or not, this is a rather arbitrary choice to me, with advantages (such as requiring a file `Inductive.v`) and drawbacks (like being able to define a bit questionable sentences such as `Ltac Inductive Inductive := apply Inductive` or `Inductive Inductive := CoInductive.`).