@Zimmi48 2017-05-05
@herbelin Well Coq is already so lax on many points, I don't feel like allowing `Ltac Inductive Inductive := apply Inductive` would be such a big deal.
If this is the only drawback, then I would say to @amintimany: just go ahead and add the commit to this PR.