@maximedenes 2017-05-05
@amintimany Something I'm wondering is: why is it ok to have cumulativity on instances but not on parameters?

For instance, why is the following rejected?
```coq
Set Universe Polymorphism.
Set Ind Cumulativity.

Inductive List (A: Type) := nil | cons : A -> List A -> List A.

Section Test1.
  Universe i j i' j'.

  Constraint i < j.
  Constraint i' < j'.

  Fail Definition LiftL : List@{i'} Type@{i} -> List@{j'} Type@{j} := fun x => x.

End Test1.
```

Sorry if this is a naive question.