@amintimany 2017-05-05
@maximedenes this I would say is an incompleteness. The problem is that when we do the analysis at the time of declaration of List we see a variable ```A``` and we don't consider what would happen if we have a ```B``` with ```A <= B``` in our analysis. Actually if you do ```Print List``` you see
````
Cumulative Inductive List@{Top.7} (A : Type@{Top.7}) : Type@{max(Set, Top.7)} :=
    nil : List@{Top.7} A | cons : A -> List@{Top.7} A -> List@{Top.7} A
(* Top.7 |= 
   
   ~@{Top.7} <= ~@{Top.9} iff 
    *)
````
The part ```~@{Top.7} <= ~@{Top.9} iff ``` is the set of universe constraints (empty here) of subtyping constraints for ```List@{Top.7} <= List@{Top.9}```.

Intuitively, I would say that in general this is true that if a parameter is a subtype of another parameter the subtyping relation should hold but we haven't done that.