@amintimany 2017-05-05
@silene what do you mean? What plugins actually declare "Inductive" as a keyword!?