@silene 2017-05-05
> What plugins actually declare "Inductive" as a keyword!?

Any plugin that contains a grammar rule containing the string `"Inductive"` not prefixed by `IDENT`. Example: https://github.com/coq/coq/blob/trunk/plugins/extraction/g_extraction.ml4#L149