@matejkosik 2017-06-14
benchmarking [results](https://ci.inria.fr/coq/view/benchmarking/job/benchmark-part-of-the-branch/lastBuild/console)
```
┌────────────────────────┬─────────────────────────┬─────────────────────────────────────┬───────────────────────────────────────┐
│                        │      user time [s]      │             CPU cycles              │           CPU instructions            │
│                        │                         │                                     │                                       │
│           package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │            NEW            OLD PDIFF   │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│         coq-coquelicot │   72.10   72.38 -0.39 % │  198356234868  198488264353 -0.07 % │   248754890394   250513895394 -0.70 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│                 coq-sf │   44.85   45.01 -0.36 % │  123201735854  123098206702 +0.08 % │   160310261523   160467528931 -0.10 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│           coq-compcert │  807.96  810.10 -0.26 % │ 2242267239720 2249536210542 -0.32 % │  3406266058321  3403526212867 +0.08 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│    coq-formal-topology │   37.70   37.58 +0.32 % │  101416983581  101066770653 +0.35 % │   138246311276   138295617759 -0.04 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│                coq-vst │ 1365.17 1360.54 +0.34 % │ 3792588788650 3778865503913 +0.36 % │  5698473487985  5693243423757 +0.09 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│              coq-color │  582.72  579.70 +0.52 % │ 1616645375410 1608218506123 +0.52 % │  2032350353218  2030703990275 +0.08 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│   coq-mathcomp-algebra │  170.71  169.78 +0.55 % │  474215940280  471316906085 +0.62 % │   676010239954   675185283233 +0.12 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│  coq-mathcomp-fingroup │   58.76   58.43 +0.56 % │  162243447589  161106845699 +0.71 % │   228201133984   228014652004 +0.08 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-character │  273.28  271.70 +0.58 % │  760025182353  754909532326 +0.68 % │  1118306569319  1116354781845 +0.17 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│              coq-flocq │   56.07   55.73 +0.61 % │  153564768300  153391068728 +0.11 % │   197785252508   198414509912 -0.32 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│             coq-geocoq │ 2176.77 2161.44 +0.71 % │ 6060217302071 6008924550476 +0.85 % │ 10121653725316 10084806569043 +0.37 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│     coq-mathcomp-field │  458.72  455.40 +0.73 % │ 1276206284323 1265699174491 +0.83 % │  2127801735139  2119948573590 +0.37 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-ssreflect │   37.03   36.76 +0.73 % │  101026633244  100769989576 +0.25 % │   125756786912   125827370289 -0.06 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│  coq-mathcomp-solvable │  193.69  192.18 +0.79 % │  537670703718  532871639381 +0.90 % │   783754980293   781140235526 +0.33 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│               coq-iris │  357.98  355.15 +0.80 % │  994683550849  986753999413 +0.80 % │  1300738036858  1296655566180 +0.31 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│         coq-verdi-raft │ 1351.16 1338.52 +0.94 % │ 3750703243478 3718829351461 +0.86 % │  5339043146329  5336717789846 +0.04 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│            coq-unimath │ 1177.87 1166.78 +0.95 % │ 3264977948078 3230842927678 +1.06 % │  5607452860718  5600601139206 +0.12 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┤
│ coq-mathcomp-odd_order │ 1412.26 1395.03 +1.24 % │ 3929367168565 3887436396428 +1.08 % │  6862007655153  6860921238965 +0.02 % │
└────────────────────────┴─────────────────────────┴─────────────────────────────────────┴───────────────────────────────────────┘

PDIFF = proportional difference between measurements done for the NEW and the OLD Coq version
      = (NEW_measurement - OLD_measurement) / OLD_measurement * 100%

NEW = 8e8dbebe7912cbfba1d3bfbe3fcaba71ceb625b5
OLD = 83d8b081c02cfde83c8fd93102f8f1aae3fe87b3
```